# docs/PRD/0004_API_계약/openapi3.json
{
  "components" : {
    "schemas" : {
      "AddCartItemRequest" : {
        "properties" : {
          "productId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "qty" : {
            "format" : "int32",
            "maximum" : 99,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "productId" ],
        "type" : "object"
      },
...
      "CaptureClientEventRequest" : {
        "properties" : {
          "event" : {
            "enum" : [ "GUEST_LOOKUP_HUB_VIEWED", "GUEST_ORDER_DIRECT_ENTRY_CONTINUED", "GUEST_MEMBER_CTA_CLICKED", "GUEST_CLAIM_MODAL_OPENED", "GUEST_CLAIM_COMPLETED" ],
            "type" : "string"
          },
          "path" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "source" : {
            "maxLength" : 80,
            "minLength" : 0,
            "type" : "string"
          },
          "target" : {
            "maxLength" : 80,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "event", "path" ],
        "type" : "object"
      },
...
      "MergeCartItemRequest" : {
        "properties" : {
          "productId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "qty" : {
            "format" : "int32",
            "maximum" : 99,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "productId" ],
        "type" : "object"
      },
      "MergeCartRequest" : {
        "properties" : {
          "idempotencyKey" : {
            "format" : "uuid",
            "type" : "string"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/MergeCartItemRequest"
            },
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [ "idempotencyKey", "items" ],
        "type" : "object"
      },
...
      "UpdateCartItemRequest" : {
        "properties" : {
          "qty" : {
            "format" : "int32",
            "maximum" : 99,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "happyGallery API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/me/cart/items" : {
      "post" : {
        "operationId" : "addMyCartItem",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddCartItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        },
        "tags" : [ "me-cart-controller" ]
      }
    },
    "/api/v1/me/cart/items/{productId}" : {
      "delete" : {
        "operationId" : "removeMyCartItem",
        "parameters" : [ {
          "in" : "path",
          "name" : "productId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "me-cart-controller" ]
      },
      "put" : {
        "operationId" : "updateMyCartItemQuantity",
        "parameters" : [ {
          "in" : "path",
          "name" : "productId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateCartItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "me-cart-controller" ]
      }
    },
    "/api/v1/me/cart/merge" : {
      "post" : {
        "operationId" : "mergeMyCartItems",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MergeCartRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "me-cart-controller" ]
      }
    },
...
    "/api/v1/monitoring/client-events" : {
      "post" : {
        "operationId" : "captureClientEvent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaptureClientEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "client-monitoring-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Same-origin API",
    "url" : "/"
  } ]
}